Pass BrowserContextJavaScriptEnabledTests
related to #489
